fix(auth): add demo activities for email and phone authentication with customizable UI
