+++ app/src/main/AndroidManifest.xml
@@ -55,6 +55,30 @@
             android:label="Custom Slots &amp; Theming Demo"
             android:exported="false"
             android:theme="@style/Theme.FirebaseUIAndroid" />
+
+        <activity
+            android:name=".EmailAuthSlotDemoActivity"
+            android:label="Email Auth — Custom Slot"
+            android:exported="false"
+            android:theme="@style/Theme.FirebaseUIAndroid" />
+
+        <activity
+            android:name=".PhoneAuthSlotDemoActivity"
+            android:label="Phone Auth — Custom Slot"
+            android:exported="false"
+            android:theme="@style/Theme.FirebaseUIAndroid" />
+
+        <activity
+            android:name=".ShapeCustomizationDemoActivity"
+            android:label="Shape Customization"
+            android:exported="false"
+            android:theme="@style/Theme.FirebaseUIAndroid" />
+
+        <activity
+            android:name=".CustomMethodPickerDemoActivity"
+            android:label="Custom Method Picker Layout"
+            android:exported="false"
+            android:theme="@style/Theme.FirebaseUIAndroid" />
     </application>
 
 </manifest>
